Rearrange benches
