+++ Cargo.toml
@@ -10,14 +10,6 @@ repository = "https://github.com/pingcap/rust-prometheus"
 homepage = "https://github.com/pingcap/rust-prometheus"
 documentation = "https://docs.rs/prometheus"
 
-include = [
-    "LICENSE",
-    "Cargo.toml",
-    "src/**/*.rs",
-    "proto/**/*.rs",
-    "benches/**/*.rs",
-]
-
 [badges]
 travis-ci = { repository = "pingcap/rust-prometheus" }
 
@@ -27,10 +19,6 @@ dev = ["clippy"]
 nightly = ["libc", "spin/unstable"]
 push = ["hyper", "libc"]
 process = ["libc", "procinfo"]
-
-[[bench]]
-name = "benches"
-path = "benches/benches.rs"
 
 [dependencies]
 protobuf = "=1.4.0"
