+++ benches/histogram.rs
@@ -11,6 +11,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#![feature(test)]
+
+extern crate prometheus;
+extern crate test;
+
 use prometheus::{Histogram, HistogramOpts, HistogramVec};
 use test::Bencher;
